+++ client-server/circularOrderedArrayTestCase.c
@@ -1,5 +1,6 @@
 #include "circularOrderedArray.h"
 #include <stdio.h>
+#include <inttypes.h>
 
 int successTests = 0;
 int failedTests = 0;
@@ -17,17 +18,17 @@ void assert(char* testName, int actual, int expected) {
 	
 }
 
-void assertPositions(char* testName, double * arrayActual, double * arrayExpected, int size){
+void assertPositions(char* testName, int64_t * arrayActual, int64_t * arrayExpected, int size){
 
 	int fail = 0;
 	
 	printf("Test: %s. ", testName);	
 	for (int i = 0; i < size; i++){
 		if(arrayActual[i] == arrayExpected[i]){
-			printf(" %f✓", arrayActual[i]);	
+			printf(" %"PRId64"✓", arrayActual[i]);	
 		} else {
 			fail = 1;
-			printf(" %f!=%f", arrayActual[i], arrayExpected[i]);	
+			printf(" %"PRId64"!=%" PRId64, arrayActual[i], arrayExpected[i]);	
 		}
 	}
 
@@ -52,8 +53,8 @@ void correctInsertStarting() {
 	assert("0. nextPos", array.next, 0);
 
 	insertOrdered(&array, 2);
-	double fifo[] = {2};
-	double ordered[] = {2};
+	int64_t fifo[] = {2};
+	int64_t ordered[] = {2};
 	assert("1. size", array.size, 1);
 	assert("1. nextPos", array.next, 0);	
 	assertPositions("1. fifo", array.fifo, fifo, 1);
@@ -61,8 +62,8 @@ void correctInsertStarting() {
 	assert("1. median", median(&array), 2);
 
 	insertOrdered(&array, 4);
-	double fifo2[] = {2, 4};
-	double ordered2[] = {2, 4};
+	int64_t fifo2[] = {2, 4};
+	int64_t ordered2[] = {2, 4};
 	assert("2. size", array.size, 2);
 	assert("2. nextPos", array.next, 0);	
 	assertPositions("2. fifo", array.fifo, fifo2, 2);
@@ -70,8 +71,8 @@ void correctInsertStarting() {
 	assert("2. median", median(&array), 4);
 
 	insertOrdered(&array, 3);
-	double fifo3[] = {2, 4, 3};
-	double ordered3[] = {2, 3, 4};
+	int64_t fifo3[] = {2, 4, 3};
+	int64_t ordered3[] = {2, 3, 4};
 	assert("3. size", array.size, 3);
 	assert("3. nextPos", array.next, 0);	
 	assertPositions("3. fifo", array.fifo, fifo3, 3);
@@ -79,8 +80,8 @@ void correctInsertStarting() {
 	assert("3. median", median(&array), 3);
 
 	insertOrdered(&array, 0);
-	double fifo4[] = {2, 4, 3, 0};
-	double ordered4[] = {0, 2, 3, 4};
+	int64_t fifo4[] = {2, 4, 3, 0};
+	int64_t ordered4[] = {0, 2, 3, 4};
 	assert("4. size", array.size, 4);
 	assert("4. nextPos", array.next, 0);	
 	assertPositions("4. fifo", array.fifo, fifo4, 4);
@@ -97,8 +98,8 @@ void correctInsertOnceFull() {
 		insertOrdered(&array, i);
 	}
 	
-	double fifo_0[] = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
-	double ordered_0[] = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
+	int64_t fifo_0[] = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
+	int64_t ordered_0[] = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
 	assert("_0. size", array.size, 10);
 	assert("_0. nextPos", array.next, 0);
 	assertPositions("_0. fifo", array.fifo, fifo_0, 10);
@@ -106,8 +107,8 @@ void correctInsertOnceFull() {
 	assert("_0. median", median(&array), 5);
 
 	insertOrdered(&array, 2);
-	double fifo_1[] = {2, 1, 2, 3, 4, 5, 6, 7, 8, 9};
-	double ordered_1[] = {1, 2, 2, 3, 4, 5, 6, 7, 8, 9};
+	int64_t fifo_1[] = {2, 1, 2, 3, 4, 5, 6, 7, 8, 9};
+	int64_t ordered_1[] = {1, 2, 2, 3, 4, 5, 6, 7, 8, 9};
 	assert("_1. size", array.size, 10);
 	assert("_1. nextPos", array.next, 1);
 	assertPositions("_1. fifo", array.fifo, fifo_1, 10);
@@ -115,8 +116,8 @@ void correctInsertOnceFull() {
 	assert("_1. median", median(&array), 5);
 
 	insertOrdered(&array, 8);
-	double fifo_2[] = {2, 8, 2, 3, 4, 5, 6, 7, 8, 9};
-	double ordered_2[] = {2, 2, 3, 4, 5, 6, 7, 8, 8, 9};
+	int64_t fifo_2[] = {2, 8, 2, 3, 4, 5, 6, 7, 8, 9};
+	int64_t ordered_2[] = {2, 2, 3, 4, 5, 6, 7, 8, 8, 9};
 	assert("_2. size", array.size, 10);
 	assert("_2. nextPos", array.next, 2);
 	assertPositions("_2. fifo", array.fifo, fifo_2, 10);
@@ -126,8 +127,8 @@ void correctInsertOnceFull() {
 	insertOrdered(&array, -1);
 	insertOrdered(&array, 1);
 	insertOrdered(&array, 0);
-	double fifo_3[] = {2, 8, -1, 1, 0, 5, 6, 7, 8, 9};
-	double ordered_3[] = {-1, 0, 1, 2, 5, 6, 7, 8, 8, 9};
+	int64_t fifo_3[] = {2, 8, -1, 1, 0, 5, 6, 7, 8, 9};
+	int64_t ordered_3[] = {-1, 0, 1, 2, 5, 6, 7, 8, 8, 9};
 	assert("_3. size", array.size, 10);
 	assert("_3. nextPos", array.next, 5);
 	assertPositions("_3. fifo", array.fifo, fifo_3, 10);
